+++ media/Changelog.md
@@ -12,6 +12,7 @@
 
 - Improve ordering editors
 - Update locale files
+- Migration for gnome 48
 
 ## New Features
 
@@ -34,7 +35,6 @@
 - Fix #170, some quick toggles are not hiding
 - Fix 'has been already disposed' error on weather widget
 - Fix '(intermediate value).Extension.features is null' error when extension unloading
-- Fix 'gschemas.compiled' not found error
 
 # 2.1-pre6
 <!-- @BuildNumber: 6 -->